requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
Flask==2.0.1
idna==3.2
iniconfig==1.1.1
ipython-genutils==0.2.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.26
numpy==1.21.0
packaging==21.0
pandas==1.3.0
parso==0.8.2
pluggy==0.13.1
protobuf==3.17.3
py==1.10.0
pycodestyle==2.7.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
pytz==2021.1
requests==2.26.0
six==1.16.0
toml==0.10.2
traitlets==5.0.5
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==2.0.1
